Completed Java E-Commerce Web Application

+++ devops-git/ecommerce-app/src/main/java/com/ecommerce/model/Product.java
@@ -4,12 +4,12 @@ public class Product {
 
     private int id;
     private String name;
-    private int price;
+    private double price;
 
     public Product() {
     }
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -31,11 +31,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
